requirements.txt
asgiref==3.6.0
coverage==7.1.0
dj-database-url==1.3.0
Django==4.1.5
django-cors-headers==3.14.0
django-countries==7.5.1
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.6
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
whitenoise==6.4.0
